requirements.txt
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
dnspython==1.16.0
Flask==1.1.1
Flask-PyMongo==2.3.0
heroku==0.1.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pymongo==3.9.0
python-dateutil==1.5
python-dotenv==0.10.3
requests==2.22.0
soupsieve==1.9.4
urllib3==1.25.6
Werkzeug==0.16.0
WTForms==2.2.1